Moved around code, and fixed bug with PW
