+++ PokeDex/app/src/main/java/com/minneydev/pokedex/util/PokemonManager.kt
@@ -1,11 +1,15 @@
 package com.minneydev.pokedex.util
 
+import androidx.work.*
 import com.minneydev.pokedex.App
 import com.minneydev.pokedex.model.pokemon.ApiPokemon
 import com.minneydev.pokedex.model.pokemon.Pokemon
+import com.minneydev.pokedex.worker.DownloadPokemonWorker
+import com.minneydev.pokedex.worker.RefreshPokemonWorker
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
+import java.util.concurrent.TimeUnit
 
 class PokemonManager {
 
@@ -30,6 +34,41 @@ class PokemonManager {
             4 -> "387"
             else -> "0"
         }
+    }
+
+    fun startPeriodicRefresh() {
+        val constraints = Constraints.Builder()
+            .setRequiresBatteryNotLow(true)
+            .setRequiredNetworkType(NetworkType.NOT_ROAMING)
+            .setRequiresStorageNotLow(true)
+            .build()
+
+        val periodicRefresh = PeriodicWorkRequestBuilder<RefreshPokemonWorker>(
+            1, TimeUnit.HOURS, //Minimum of 15 min apparently
+        30, TimeUnit.MINUTES) //So then it doesn't run on the initial download
+            .setConstraints(constraints)
+            .build()
+
+        val refreshManager = WorkManager.getInstance(App.getAppContext())
+        refreshManager.enqueueUniquePeriodicWork("REFRESH",
+            ExistingPeriodicWorkPolicy.REPLACE, periodicRefresh)
+
+    }
+
+    fun downloadPokemon() {
+        val constraints = Constraints.Builder()
+            .setRequiresBatteryNotLow(true)
+            .setRequiresStorageNotLow(true)
+            .setRequiredNetworkType(NetworkType.NOT_ROAMING)
+            .build()
+
+        val downloadRequest = OneTimeWorkRequestBuilder<DownloadPokemonWorker>()
+            .setConstraints(constraints)
+            .build()
+
+        val workManager = WorkManager.getInstance(App.getAppContext())
+        workManager.enqueueUniqueWork("DOWNLOAD",
+            ExistingWorkPolicy.APPEND, downloadRequest)
     }
 
 }
